Separate all metadata sub-page in input panel to file
