+++ frontend/src/pages/inputPanel.tsx
@@ -1,21 +1,18 @@
-// React
-import { ChangeEvent, useState } from "react";
-
 // React Component
-import { Button, Col, Form, Row } from "react-bootstrap";
+import { Button } from "react-bootstrap";
 import Tab from "react-bootstrap/Tab";
 import Tabs from "react-bootstrap/Tabs";
 
 // Project Specified Component
-import { AgeRatingSelect, MangaSelect } from "../components/EnumSelect";
-import { TagsArea } from "../components/Tags";
 import { basename } from "../filename";
-import FormDateRow from "../forms/FormDateRow";
 import FormRow from "../forms/FormRow";
-import FormSelectRow from "../forms/FormSelectRow";
-import OptionFormRow from "../forms/OptionFormRow";
+import BookMetadata from "./metadata/BookMetadata";
+import CreatorMetadata from "./metadata/CreatorMetadata";
+import MiscMetadata from "./metadata/MiscMetadata";
+import SeriesMetadata from "./metadata/SeriesMetadata";
+import TagMetadata from "./metadata/TagMetadata";
 
-import { GetAllGenreInput } from "../../wailsjs/go/main/App";
+// Wails binding
 import { comicinfo } from "../../wailsjs/go/models";
 
 /** Props Interface for InputPanel */
@@ -33,362 +30,11 @@ type InputProps = {
 	infoSetter: (field: string, value: string | number) => void;
 };
 
-/** The Props for Metadata Component. */
-type MetadataProps = {
-	/** The comic info object. Accept undefined value. */
-	comicInfo: comicinfo.ComicInfo | undefined;
-
-	/** The info Setter. This function should be doing setting value, but no verification. */
-	infoSetter: (field: string, value: string | number) => void;
-};
-
-/**
- * The interface for show/edit book metadata.
- * @returns JSX Element
- */
-function BookMetadata({ comicInfo: info, infoSetter }: Readonly<MetadataProps>) {
-	/**
-	 * Handler for all input field in this panel.
-	 * This method will use <code>infoSetter</code> as core,
-	 * and apply change to comicInfo content.
-	 * <p>
-	 * This method will try to find the field is number input first,
-	 * if field name is number/related, then it will call Number() method before set to ComicInfo
-	 *
-	 * @param e the event object, for identify target element
-	 * @returns void
-	 */
-	function handleChanges(e: ChangeEvent<HTMLInputElement> | ChangeEvent<HTMLSelectElement>) {
-		console.log(e.target.title, e.target.value, e.target.type);
-
-		// Identify Number & Year
-		if (e.target.type === "number" || e.target.type === "Year") {
-			infoSetter(e.target.title, Number(e.target.value));
-			return;
-		}
-
-		// Identify Month & Day
-		if (e.target.title === "Month" || e.target.title === "Day") {
-			infoSetter(e.target.title, Number(e.target.value));
-			return;
-		}
-
-		// Normal Cases
-		infoSetter(e.target.title, e.target.value);
-	}
-
-	return (
-		<div>
-			<Form>
-				<FormRow title={"Title"} value={info?.Title} onChange={handleChanges} />
-				<FormRow title={"Summary"} value={info?.Summary.InnerXML} textareaRow={3} onChange={handleChanges} />
-				<FormRow title={"Number"} value={info?.Number} onChange={handleChanges} />
-				<FormDateRow
-					title={"Year/Month/Day"}
-					year={info?.Year}
-					month={info?.Month}
-					day={info?.Day}
-					onYearChange={handleChanges}
-					onSelectChange={handleChanges}
-				/>
-				<FormRow title={"Web"} value={info?.Web} onChange={handleChanges} />
-				<FormRow title={"GTIN"} value={info?.GTIN} onChange={handleChanges} />
-			</Form>
-		</div>
-	);
-}
-
-/**
- * The interface for show/edit creator metadata.
- * @returns JSX Element
- */
-function CreatorMetadata({ comicInfo: info, infoSetter }: Readonly<MetadataProps>) {
-	/**
-	 * Handler for all input field in this panel.
-	 * This method will use <code>infoSetter</code> as core,
-	 * and apply change to comicInfo content.
-	 * <p>
-	 * This method will try to find the field is number input first,
-	 * if field name is number/related, then it will call Number() method before set to ComicInfo
-	 *
-	 * @param e the event object, for identify target element
-	 * @returns void
-	 */
-	function handleChanges(e: ChangeEvent<HTMLInputElement> | ChangeEvent<HTMLSelectElement>) {
-		console.log(e.target.title, e.target.value, e.target.type);
-
-		// Identify Number & Year
-		if (e.target.type === "number" || e.target.type === "Year") {
-			infoSetter(e.target.title, Number(e.target.value));
-			return;
-		}
-
-		// Identify Month & Day
-		if (e.target.title === "Month" || e.target.title === "Day") {
-			infoSetter(e.target.title, Number(e.target.value));
-			return;
-		}
-
-		// Normal Cases
-		infoSetter(e.target.title, e.target.value);
-	}
-
-	return (
-		<div>
-			<Form>
-				<FormRow title={"Writer"} value={info?.Writer} onChange={handleChanges} />
-				<FormRow title={"Penciller"} value={info?.Penciller} onChange={handleChanges} />
-				<FormRow title={"Inker"} value={info?.Inker} onChange={handleChanges} />
-				<FormRow title={"Colorist"} value={info?.Colorist} onChange={handleChanges} />
-				<FormRow title={"Letterer"} value={info?.Letterer} onChange={handleChanges} />
-				<FormRow title={"CoverArtist"} value={info?.CoverArtist} onChange={handleChanges} />
-				<FormRow title={"Editor"} value={info?.Editor} onChange={handleChanges} />
-				<FormRow title={"Translator"} value={info?.Translator} onChange={handleChanges} />
-				<FormRow title={"Publisher"} value={info?.Publisher} onChange={handleChanges} />
-			</Form>
-		</div>
-	);
-}
-
-/** The user interface for show/edit Series MetaData. */
-function SeriesMetadata({ comicInfo, infoSetter }: Readonly<MetadataProps>) {
-	/**
-	 * Handler for all input field in this panel.
-	 * This method will use <code>infoSetter</code> as core,
-	 * and apply change to comicInfo content.
-	 * <p>
-	 * This method will try to find the field is number input first,
-	 * if field name is number/related, then it will call Number() method before set to ComicInfo
-	 *
-	 * @param e the event object, for identify target element
-	 * @returns void
-	 */
-	function handleChanges(e: ChangeEvent<HTMLInputElement> | ChangeEvent<HTMLSelectElement>) {
-		console.log(e.target.title, e.target.value, e.target.type);
-
-		// Identify Number
-		if (e.target.type === "number") {
-			infoSetter(e.target.title, Number(e.target.value));
-			return;
-		}
-
-		// Normal Cases
-		infoSetter(e.target.title, e.target.value);
-	}
-
-	return (
-		<div>
-			<Form>
-				<FormRow title={"Series"} value={comicInfo?.Series} onChange={handleChanges} />
-				<FormRow title={"Volume"} value={comicInfo?.Volume} onChange={handleChanges} />
-				<FormRow title={"Count"} value={comicInfo?.Count} onChange={handleChanges} />
-				<FormSelectRow
-					title={"AgeRating"}
-					selectElement={<AgeRatingSelect value={comicInfo?.AgeRating} onChange={handleChanges} />}
-				/>
-				<FormSelectRow
-					title={"Manga"}
-					selectElement={<MangaSelect value={comicInfo?.Manga} onChange={handleChanges} />}
-				/>
-				<OptionFormRow
-					title={"Genre"}
-					value={comicInfo?.Genre}
-					setValue={(val) => infoSetter("Genre", val)}
-					getDefaultOpt={GetAllGenreInput}
-				/>
-				<FormRow title={"LanguageISO"} value={comicInfo?.LanguageISO} onChange={handleChanges} />
-			</Form>
-		</div>
-	);
-}
-
-/** The user interface for show/edit Collection & ReadList MetaData. */
-function MiscMetadata({ comicInfo, infoSetter }: Readonly<MetadataProps>) {
-	/**
-	 * Handler for all input field in this panel.
-	 * This method will use <code>infoSetter</code> as core,
-	 * and apply change to comicInfo content.
-	 * <p>
-	 * This method will try to find the field is number input first,
-	 * if field name is number/related, then it will call Number() method before set to ComicInfo
-	 *
-	 * @param e the event object, for identify target element
-	 * @returns void
-	 */
-	function handleChanges(e: ChangeEvent<HTMLInputElement> | ChangeEvent<HTMLSelectElement>) {
-		console.log(e.target.title, e.target.value, e.target.type);
-
-		// Identify Number & Year
-		if (e.target.type === "number" || e.target.type === "Year") {
-			infoSetter(e.target.title, Number(e.target.value));
-			return;
-		}
-
-		// Identify Month & Day
-		if (e.target.title === "Month" || e.target.title === "Day") {
-			infoSetter(e.target.title, Number(e.target.value));
-			return;
-		}
-
-		// Normal Cases
-		infoSetter(e.target.title, e.target.value);
-	}
-
-	return (
-		<div>
-			<Form>
-				<FormRow title={"SeriesGroup"} value={comicInfo?.SeriesGroup} onChange={handleChanges} />
-				<FormRow title={"AlternateSeries"} value={comicInfo?.AlternateSeries} onChange={handleChanges} />
-				<FormRow title={"AlternateNumber "} value={comicInfo?.AlternateNumber} onChange={handleChanges} />
-				<FormRow title={"AlternateCount"} value={comicInfo?.AlternateCount} onChange={handleChanges} />
-				<FormRow title={"StoryArc"} value={comicInfo?.StoryArc} onChange={handleChanges} />
-				<FormRow title={"StoryArcNumber"} value={comicInfo?.StoryArcNumber} onChange={handleChanges} />
-			</Form>
-		</div>
-	);
-}
-
-/**
- * The interface for show/edit tags metadata.
- * @returns JSX Element
- */
-function TagMetadata({ comicInfo: info, infoSetter }: Readonly<MetadataProps>) {
-	/** Hooks of tag that to be added. Only allow single tag to be added at a time. */
-	const [singleTag, setSingleTag] = useState<string>("");
-
-	/** Function for handing enter key pressed when entering custom tags. */
-	const handleKeyDown = (event: React.KeyboardEvent) => {
-		if (event.key === "Enter") {
-			event.preventDefault();
-			handleAdd();
-		}
-	};
-
-	/**
-	 * Function to handle the textfield of tag to be added.
-	 * @param e the react event
-	 */
-	function handleChanges(e: ChangeEvent<HTMLInputElement>) {
-		setSingleTag(e.target.value);
-	}
-
-	/** Function for handling add button click */
-	function handleAdd() {
-		// Prevent Empty tags
-		if (singleTag === "") {
-			return;
-		}
-
-		// Retrieve Tags
-		let temp = info?.Tags;
-
-		// Append Tags to comic info
-		if (temp === "" || temp === undefined) {
-			temp = singleTag;
-		} else {
-			temp += ", " + singleTag;
-		}
-
-		// Apply Change to ComicInfo
-		infoSetter("Tags", temp);
-
-		// Empty Tags in textfield
-		setSingleTag("");
-	}
-
-	/** Function for handling delete button click */
-	function handleDelete(id: number) {
-		if (info?.Tags === undefined) {
-			return;
-		}
-
-		// Parse tags to array of strings
-		let array = info.Tags.split(",");
-
-		// Remove tag by index
-		array.splice(id, 1);
-
-		// Concat array of string to string
-		let str = array.join(", ");
-
-		// Set by info setter
-		infoSetter("Tags", str);
-	}
-
-	return (
-		<div>
-			<Form>
-				{/* A Text Area for holding lines of tags. */}
-				<Row>
-					<Col sm={2} className="mt-1">
-						{"Tags"}
-					</Col>
-					<Col sm={9}>
-						<TagsArea rawTags={info?.Tags} handleDelete={handleDelete} />
-					</Col>
-				</Row>
-
-				{/* Empty Rows for margin */}
-				<Row className="mb-3" />
-
-				<Row>
-					{/* Empty Column */}
-					<Col sm={2} className="mt-1">
-						Add Custom Tag
-					</Col>
-
-					{/* Column of adding tags */}
-					<Col sm={8}>
-						<Form.Control value={singleTag} onChange={handleChanges} onKeyDown={handleKeyDown} />
-					</Col>
-
-					{/* Column of add button */}
-					<Col sm={1}>
-						<Button variant="outline-info" onClick={handleAdd}>
-							Add
-						</Button>
-					</Col>
-				</Row>
-			</Form>
-		</div>
-	);
-}
-
 /**
  * The panel for input/edit content of ComicInfo.xml
  * @returns JSX Element
  */
 export default function InputPanel({ comicInfo, folderName, exportFunc, infoSetter }: Readonly<InputProps>) {
-	/**
-	 * Handler for all input field in this panel.
-	 * This method will use <code>infoSetter</code> as core,
-	 * and apply change to comicInfo content.
-	 * <p>
-	 * This method will try to find the field is number input first,
-	 * if field name is number/related, then it will call Number() method before set to ComicInfo
-	 *
-	 * @param e the event object, for identify target element
-	 * @returns void
-	 */
-	function handleChanges(e: ChangeEvent<HTMLInputElement> | ChangeEvent<HTMLSelectElement>) {
-		console.log(e.target.title, e.target.value, e.target.type);
-
-		// Identify Number & Year
-		if (e.target.type === "number" || e.target.type === "Year") {
-			infoSetter(e.target.title, Number(e.target.value));
-			return;
-		}
-
-		// Identify Month & Day
-		if (e.target.title === "Month" || e.target.title === "Day") {
-			infoSetter(e.target.title, Number(e.target.value));
-			return;
-		}
-
-		// Normal Cases
-		infoSetter(e.target.title, e.target.value);
-	}
-
 	return (
 		<div id="Input-Panel" className="mt-5">
 			<h5 className="mb-4">Modify ComicInfo.xml</h5>
